requirements.txt
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.3.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.3
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
cryptography==44.0.2
dash==3.0.2
dateparser==1.2.1
Deprecated==1.2.18
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.0.3
fsspec==2025.3.2
graphviz==0.20.3
greenlet==3.1.1
griffe==1.7.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.33.0
nest-asyncio==1.6.0
numpy==2.2.4
oauthlib==3.2.2
opentelemetry-api==1.31.1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
plotly==6.0.1
prefect==3.3.3
prometheus_client==0.21.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.2
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.3.0
uv==0.6.12
uvicorn==0.34.0
websockets==15.0.1
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.21.0
